api/dev.requirements.txt
annotated-types==0.7.0
anyio==4.4.0
certifi==2024.6.2
charset-normalizer==3.4.1
click==8.1.7
contourpy==1.3.1
cycler==0.12.1
dnspython==2.6.1
email_validator==2.2.0
face-detection @ git+https://github.com/elliottzheng/face-detection.git@786fbab7095623c348e251f1f0a8b323721c6a84
fastapi==0.111.0
fastapi-cli==0.0.4
filelock==3.16.1
fonttools==4.55.0
fsspec==2024.10.0
graphql-core==3.2.3
gunicorn==23.0.0
h11==0.14.0
httpcore==1.0.5
httptools==0.6.1
httpx==0.27.0
huggingface-hub==0.27.1
idna==3.7
Jinja2==3.1.4
kiwisolver==1.4.7
markdown-it-py==3.0.0
MarkupSafe==2.1.5
matplotlib==3.9.2
mdurl==0.1.2
mpmath==1.3.0
networkx==3.4.2
numpy==2.0.0
opencv-python==4.10.0.84
orjson==3.10.5
packaging==24.2
pandas==2.2.3
pillow==11.0.0
pydantic==2.7.4
pydantic_core==2.18.4
Pygments==2.18.0
PyJWT==2.10.1
PyMySQL==1.1.1
pyparsing==3.2.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-multipart==0.0.9
pytz==2024.2
PyYAML==6.0.1
requests==2.32.3
rich==13.7.1
scipy==1.14.1
shellingham==1.5.4
six==1.16.0
sixdrepnet==0.1.6
sniffio==1.3.1
SQLAlchemy==2.0.31
starlette==0.37.2
strawberry-graphql==0.235.1
sympy==1.13.1
torch==2.5.1
torchvision==0.20.1
tqdm==4.67.1
typer==0.12.3
typing_extensions==4.12.2
tzdata==2024.2
ujson==5.10.0
urllib3==2.3.0
uvicorn==0.30.1
uvloop==0.19.0
watchfiles==0.22.0
websockets==12.0
